Refactor and relocate packages from pkg/ that were not actually common
- All test variations pass.
