+++ internal/accumulator/accumulator/accumulator.go
@@ -12,8 +12,8 @@ import (
 	"github.com/thingspect/api/go/common"
 	"github.com/thingspect/atlas/internal/accumulator/config"
 	"github.com/thingspect/atlas/pkg/alog"
+	"github.com/thingspect/atlas/pkg/dao"
 	"github.com/thingspect/atlas/pkg/dao/datapoint"
-	"github.com/thingspect/atlas/pkg/postgres"
 	"github.com/thingspect/atlas/pkg/queue"
 )
 
@@ -37,7 +37,7 @@ type Accumulator struct {
 // value.
 func New(cfg *config.Config) (*Accumulator, error) {
 	// Set up database connection.
-	pg, err := postgres.New(cfg.PgURI)
+	pg, err := dao.NewPgDB(cfg.PgURI)
 	if err != nil {
 		return nil, err
 	}
