+++ internal/api/crypto/constants.go
