+++ internal/api/crypto/encrypt.go
