+++ internal/api/lora/fake.go
