+++ internal/api/service/device_test.go
@@ -12,9 +12,9 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/thingspect/api/go/api"
 	"github.com/thingspect/api/go/common"
+	"github.com/thingspect/atlas/internal/api/lora"
 	"github.com/thingspect/atlas/internal/api/session"
 	"github.com/thingspect/atlas/pkg/dao"
-	"github.com/thingspect/atlas/pkg/lora"
 	"github.com/thingspect/atlas/pkg/test/matcher"
 	"github.com/thingspect/atlas/pkg/test/random"
 	"google.golang.org/grpc/codes"
