+++ internal/api/service/user_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/thingspect/api/go/api"
 	"github.com/thingspect/api/go/common"
+	"github.com/thingspect/atlas/internal/api/crypto"
 	"github.com/thingspect/atlas/internal/api/session"
-	"github.com/thingspect/atlas/pkg/crypto"
 	"github.com/thingspect/atlas/pkg/dao"
 	"github.com/thingspect/atlas/pkg/test/matcher"
 	"github.com/thingspect/atlas/pkg/test/random"
