+++ tool/atlas-hash/main.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/thingspect/atlas/internal/api/crypto"
 	"github.com/thingspect/atlas/internal/api/session"
-	"github.com/thingspect/atlas/pkg/crypto"
 )
 
 const usage = `Usage:
